Added full authentication
Log In, Log Out, authGuard functions (requiresAuth, requiresRole) and errors. Generate JWTs and store them as httpOnly cookies, server signs them with Config.SECRET_KEY as key, and verifies them with the default algorithm. For logging out, at least for the time being, the cookie is removed, in a future a blacklist could be implemented.